Added function to remove shortcut
The TeamViewer installer sometimes disregards the DESKTOPSHORTCUTS=0 parameter, resulting in the creation of a desktop shortcut. The newly added function, Remove-TeamViewerShortcut, addresses this by scanning both the Public and User desktop directories and removing the shortcut if it is found.# Please enter the commit message for your changes. Lines starting